Added gRPC service for discounts and updated dependencies
In this commit, we have integrated a gRPC service, `DiscountGrpcService`, to fetch and apply discounts to the basket items. The `Basket.API` and `Basket.Application` projects now have dependencies on `Basket.Application` and `Discount.Application` respectively. The `DiscountGrpcService` has been added as a dependency in the `BasketController` and the `Startup` class. A loop was added in the `UpdateBasket` method to apply discounts using the `DiscountGrpcService`. The `GrpcSettings` section was added to `appsettings.json` to configure the gRPC service URL. The `Grpc.Tools` package and a protobuf file `discount.proto` were added to `Basket.Application.csproj`. Lastly, `Grpc.AspNetCore` package was added to `Basket.Infrastructure.csproj` to support building gRPC services with ASP.NET Core.